CNN_for_disease_detection/requirements.txt
tensorflow==2.17.1
numpy==1.24.3
scikit-learn==1.3.2
flask
pillow